requirements.txt
alike==0.1.0
anndata==0.12.0
anyio==4.9.0
cellpose==3.1.1.1
click==8.1.7
conch==8.1.0
cucim==23.10.0
cupy==13.5.1
dask==2024.11.2
dask_image==2024.5.3
dask_jobqueue==0.9.0
deeperhistreg==1.0.1
docstring_parser==0.16
einops==0.8.1
einops_exts==0.0.4
environs==14.2.0
faiss_cpu==1.9.0
flash_attn==2.8.1
Flask==3.1.1
gdown==5.2.0
geopandas==1.0.1
h5py==3.12.1
huggingface_hub==0.27.1
imageio==2.36.0
ipython==8.12.3
kornia==0.7.4
langchain==0.3.26
langchain_core==0.3.69
langchain_mcp_adapters==0.1.9
langchain_modelscope==0.1.1
langchain_openai==0.3.28
langgraph==0.5.3
lazyslide==0.7.2
loguru==0.7.2
lovely_tensors==0.1.18
madeleine==0.1.0
magicgui==0.10.1
matlab==0.1
matplotlib==3.10.3
mcp==1.12.0
miso==3.1.25
musk==0.0.1
napari==0.5.4
nest_asyncio==1.6.0
networkx==3.4.2
numpy==2.3.1
openai==1.97.0
opencv_python==4.7.0.72
opencv_python_headless==4.10.0.84
openslide_bin==4.0.0.8
optuna==4.3.0
pandas==2.3.1
Pillow==11.3.0
playwright==1.52.0
pooch==1.8.2
poselib==2.0.4
psutil==6.1.0
pydantic==2.11.7
pykan==0.2.8
pylibCZIrw==5.0.0
pyseq==0.9.0
pytest==8.3.3
python-dotenv==1.1.1
python_igraph==0.11.9
pytorch_lightning==1.9.4
pytorch_msssim==1.0.0
pyvips==2.2.3
pyvista==0.44.1
PyYAML==6.0.2
PyYAML==6.0.2
QtPy==2.4.2
rasterio==1.4.2
Requests==2.32.4
sacremoses==0.1.1
scanpy==1.11.3
scikit_learn==1.7.1
scipy==1.16.0
seaborn==0.13.2
segmentation_models_pytorch==0.5.0
setuptools==57.5.0
Shapely==2.1.1
skimage==0.0
spatialdata==0.4.0
Sphinx==8.2.3
starlette==0.47.1
streamlit==1.45.0
tifffile==2024.9.20
timm==1.0.17
timm_ctp==0.5.4
torch==2.2.0
torchvision==0.17.0
tqdm==4.67.0
transformers==4.48.1
trimesh==4.5.2
typing_extensions==4.14.1
umap_learn==0.5.7
uvicorn==0.35.0
valis==0.1
valis_hest==0.0.1
wandb==0.21.0
webdataset==1.0.2
wsidata==0.7.0
xarray==2024.11.0
xformers==0.0.31.post1
